use long double to replace double

+++ src/Circle.cpp
@@ -11,7 +11,7 @@ Circle::Circle(const string& input) {
     this->Circle::Circle(stod(infos.at(1)), stod(infos.at(2)), stod(infos.at(3)));
 }
 
-Circle::Circle(double x, double y, double r) {
+Circle::Circle(long double x, long double y, long double r) {
     this->center = new Point(x, y);
     this->r = r;
 }
@@ -19,7 +19,7 @@ Circle::Circle(double x, double y, double r) {
 vector<Point> Circle::getIntersectionWith(Line& line) {
     vector<Point> result;
 
-    const double distance = this->center->getDistanceToLine(line);
+    const long double distance = this->center->getDistanceToLine(line);
     if (distance > this->r) {
         return result;
     }
@@ -31,12 +31,12 @@ vector<Point> Circle::getIntersectionWith(Line& line) {
     return calculateIntersectionWithNormalLine(line);
 }
 
-vector<Point> Circle::calculatePointsAtX(double x) {
+vector<Point> Circle::calculatePointsAtX(long double x) {
    vector<Point> result;
 
-   const double common_part = sqrt(pow(r, 2) - pow(x - center->x, 2));
-   const double y1 = center->y + common_part;
-   const double y2 = center->y - common_part;
+   const long double common_part = sqrt(pow(r, 2) - pow(x - center->x, 2));
+   const long double y1 = center->y + common_part;
+   const long double y2 = center->y - common_part;
 
    result.push_back(Point(x, y1));
    result.push_back(Point(x, y2));
@@ -44,23 +44,23 @@ vector<Point> Circle::calculatePointsAtX(double x) {
    return result;
 }
 
-double calulateYOfXInLine(Line& line, double x) {
+long double calulateYOfXInLine(Line& line, long double x) {
     return line.k * x + line.b;
 }
 
 vector<Point> Circle::calculateIntersectionWithNormalLine(Line& line) {
     vector<Point> result;
 
-    const double t = line.b - center->y;
+    const long double t = line.b - center->y;
 
-    const double a = pow(line.k, 2) + 1;
-    const double b = 2 * (line.k * t - center->x);
-    const double c = pow(t, 2) + pow(center->x, 2) - pow(r, 2);
+    const long double a = pow(line.k, 2) + 1;
+    const long double b = 2 * (line.k * t - center->x);
+    const long double c = pow(t, 2) + pow(center->x, 2) - pow(r, 2);
 
-    const double delta = sqrt(b * b - 4 * a * c);
+    const long double delta = sqrt(b * b - 4 * a * c);
 
-    const double x1 = (-b + delta) / (2 * a);
-    const double x2 = (-b - delta) / (2 * a);
+    const long double x1 = (-b + delta) / (2 * a);
+    const long double x2 = (-b - delta) / (2 * a);
 
     result.push_back(Point(x1, calulateYOfXInLine(line, x1)));
     result.push_back(Point(x2, calulateYOfXInLine(line, x2)));
@@ -68,7 +68,7 @@ vector<Point> Circle::calculateIntersectionWithNormalLine(Line& line) {
     return result;
 }
 
-double calculateDistanceBetweenPoints(const Point& point1, const Point& point2) {
+long double calculateDistanceBetweenPoints(const Point& point1, const Point& point2) {
     return sqrt(pow(point1.x - point2.x, 2) + pow(point1.y - point2.y, 2));
 }
 
@@ -77,9 +77,9 @@ Line getCommonLineBetweenCricles(const Circle& c1, const Circle c2) {
         return Line(INT_MAX, ((pow(c1.r, 2) - pow(c2.r, 2)) / (c2.center->x - c1.center->x) + c1.center->x + c2.center->x) / 2);
     }
 
-    const double a = 2 * (c2.center->x - c1.center->x);
-    const double b = 2 * (c2.center->y - c1.center->y);
-    const double c = (pow(c1.center->x, 2) + pow(c1.center->y, 2) - pow(c1.r, 2)) -
+    const long double a = 2 * (c2.center->x - c1.center->x);
+    const long double b = 2 * (c2.center->y - c1.center->y);
+    const long double c = (pow(c1.center->x, 2) + pow(c1.center->y, 2) - pow(c1.r, 2)) -
         (pow(c2.center->x, 2) + pow(c2.center->y, 2) - pow(c2.r, 2));
 
     return Line(-a / b, -c / b);
@@ -88,7 +88,7 @@ Line getCommonLineBetweenCricles(const Circle& c1, const Circle c2) {
 vector<Point> Circle::getIntersectionWith(Circle& another) {
     vector<Point> result;
 
-    const double distance = calculateDistanceBetweenPoints(*center, *another.center);
+    const long double distance = calculateDistanceBetweenPoints(*center, *another.center);
     
     if (distance > r + another.r || distance < abs(r - another.r)) {
         return result;
