+++ src/Point.cpp
@@ -1,7 +1,7 @@
 #include "Point.h"
 #include "Line.h"
 
-Point::Point(double x, double y) {
+Point::Point(long double x, long double y) {
 	this->x = x;
 	this->y = y;
 }
@@ -17,13 +17,13 @@ bool Point::operator<(const Point& another) const {
 	return this->x < another.x;
 }
 
-double Point::getDistanceToLine(const Line& line) {
+long double Point::getDistanceToLine(const Line& line) {
 	if (line.k == INT_MAX) {
 		return abs(this->x - line.b);
 	}
 
-	double molecule = abs(line.k * this->x + (-1) * this->y + line.b);
-	double denominator = sqrt(pow(line.k, 2) + 1);
+	long double molecule = abs(line.k * this->x + (-1) * this->y + line.b);
+	long double denominator = sqrt(pow(line.k, 2) + 1);
 
 	return molecule / denominator;
 }
